Move Tidb.h
