+++ dbms/src/Debug/dbgFuncSchema.cpp
@@ -25,7 +25,7 @@
 #include <Storages/Transaction/SchemaSyncService.h>
 #include <Storages/Transaction/SchemaSyncer.h>
 #include <Storages/Transaction/TMTContext.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 #include <fmt/core.h>
 
 #include <ext/singleton.h>
