+++ dbms/src/Debug/dbgNaturalDag.h
@@ -17,7 +17,7 @@
 #include <Flash/Coprocessor/RegionInfo.h>
 #include <Storages/Transaction/KVStore.h>
 #include <Storages/Transaction/RegionMeta.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 #include <Storages/Transaction/TiKVKeyValue.h>
 #include <kvproto/coprocessor.pb.h>
