+++ dbms/src/Debug/dbgTools.cpp
@@ -25,7 +25,7 @@
 #include <Storages/Transaction/Region.h>
 #include <Storages/Transaction/RowCodec.h>
 #include <Storages/Transaction/TMTContext.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 #include <Storages/Transaction/TiKVRange.h>
 
 #include <random>
