+++ dbms/src/Flash/Coprocessor/DAGQueryBlock.h
@@ -21,7 +21,7 @@
 
 #include <Flash/Coprocessor/DAGContext.h>
 #include <Interpreters/IQuerySource.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 #include <Storages/Transaction/TiKVKeyValue.h>
 #include <Storages/Transaction/Types.h>
