+++ dbms/src/Flash/Coprocessor/DAGQueryBlockInterpreter.cpp
@@ -51,7 +51,7 @@
 #include <Interpreters/Join.h>
 #include <Parsers/ASTSelectQuery.h>
 #include <Parsers/ASTTablesInSelectQuery.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 #include <WindowFunctions/WindowFunctionFactory.h>
