+++ dbms/src/Flash/Coprocessor/DefaultChunkCodec.cpp
@@ -15,7 +15,7 @@
 #include <Flash/Coprocessor/DefaultChunkCodec.h>
 #include <Storages/Transaction/Datum.h>
 #include <Storages/Transaction/DatumCodec.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 #include <Storages/Transaction/TypeMapping.h>
 
 using TiDB::DatumBumpy;
