+++ dbms/src/Storages/DeltaMerge/RowKeyRange.h
@@ -21,7 +21,7 @@
 #include <Storages/DeltaMerge/DeltaMergeHelpers.h>
 #include <Storages/Transaction/DatumCodec.h>
 #include <Storages/Transaction/RegionRangeKeys.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 #include <Storages/Transaction/TiKVKeyValue.h>
 #include <Storages/Transaction/TiKVRecordFormat.h>
 #include <Storages/Transaction/Types.h>
