+++ dbms/src/Storages/DeltaMerge/tests/gtest_convert_column.cpp
@@ -18,7 +18,7 @@
 #include <Storages/DeltaMerge/DeltaMergeDefines.h>
 #include <Storages/DeltaMerge/SchemaUpdate.h>
 #include <Storages/DeltaMerge/convertColumnTypeHelpers.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 
 #include "dm_basic_include.h"
