+++ dbms/src/Storages/StorageDeltaMerge.h
@@ -23,7 +23,7 @@
 #include <Storages/IManageableStorage.h>
 #include <Storages/IStorage.h>
 #include <Storages/Transaction/DecodingStorageSchemaSnapshot.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 
 #include <ext/shared_ptr_helper.h>
