+++ dbms/src/Storages/System/StorageSystemTables.cpp
@@ -28,7 +28,7 @@
 #include <Storages/MutableSupport.h>
 #include <Storages/System/StorageSystemTables.h>
 #include <Storages/Transaction/SchemaNameMapper.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 #include <Storages/Transaction/Types.h>
 #include <Storages/VirtualColumnUtils.h>
