+++ dbms/src/Storages/Transaction/Datum.h
@@ -17,7 +17,7 @@
 #include <optional>
 
 #include <Core/Field.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 
 namespace TiDB
 {
