+++ dbms/src/Storages/Transaction/DecodingStorageSchemaSnapshot.h
@@ -16,7 +16,7 @@
 
 #include <Common/nocopyable.h>
 #include <Storages/DeltaMerge/DeltaMergeDefines.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 
 
 namespace DB
