+++ dbms/src/Storages/Transaction/SchemaGetter.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wunused-parameter"
 #pragma GCC diagnostic ignored "-Wnon-virtual-dtor"
