+++ dbms/src/Storages/Transaction/SchemaNameMapper.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 
 namespace DB
 {
