+++ dbms/src/Storages/Transaction/TypeMapping.h
@@ -16,7 +16,7 @@
 
 #include <Core/NamesAndTypes.h>
 #include <DataTypes/IDataType.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wunused-parameter"
 #include <tipb/expression.pb.h>
