+++ dbms/src/Storages/Transaction/tests/gtest_type_mapping.cpp
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include <DataTypes/DataTypeFactory.h>
-#include <Storages/Transaction/TiDB.h>
+#include <src/TiDB/TiDB.h>
 #include <Storages/Transaction/TypeMapping.h>
 #include <TestUtils/TiFlashTestBasic.h>
